requirements.txt
mysqlclient
python-dotenv
PyYAML
requests